csc486b/a1/requirements.txt
appnope
backcall
cycler
decorator
h5py
ipython
ipython-genutils
jedi
kiwisolver
matplotlib
numpy
parso
pexpect
pickleshare
Pillow
prompt-toolkit
ptyprocess
Pygments
pyparsing
python-dateutil
scipy
six
traitlets
wcwidth
